+++ projects/lionheart/Plan.cpp
@@ -43,14 +43,14 @@ lionheart::Paths::Paths(std::shared_ptr<const Map> const& map, int maxSpeed) : v
     lionheart::Direction::NORTH, lionheart::Direction::EAST, lionheart::Direction::SOUTH, lionheart::Direction::WEST};
   //Create all vertexes
   auto vNum = 0;
-  for(size_t r=0;r<map->rows();++r)
-    for(size_t c=0;c<map->cols();++c)
+  for(int r=0;r<map->rows();++r)
+    for(int c=0;c<map->cols();++c)
     {
       if((*map)[map->at(r,c)] == Tile::SPACE)
       {
         for(auto&& d:dirs)
         {
-          vertex[{{static_cast<int>(r), static_cast<int>(c)}, d}] = vNum;
+          vertex[{{r, c}, d}] = vNum;
           ++vNum;
         }
       }
